Move installed_schemas fixture to .testing
